Refactoring: scheduleCard component
